+++ src/main/java/com/ada/currencycalc/model/RateDTO.java
@@ -3,10 +3,12 @@ package com.ada.currencycalc.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class RateDTO {
     private String no;
     private String effectiveDate;
-    private double mid;
+    private BigDecimal mid;
 }
